base_class_mpdrnn.py - Parent class created. - mpdrnn.py and param_search_mpdrnn.py inherits from this class
data_paths.py:
- Unnecessary and unused variables are erased

dataset_config.py:
- same as above
